Extract tests about certificate selection
